requirements.txt
streamlit==1.15.1
plotly==5.10.0
pandas==1.4.3
numpy==1.23.1
folium==0.16.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
haversine==2.8.1
streamlit-folium==0.18.0
Pillow==10.2.0
